backend/requirements.txt
Flask==2.3.3
flask-sqlalchemy==3.0.5
flask-mail==0.9.1
psycopg2-binary==2.9.9
werkzeug==2.3.7
openpyxl==3.1.2

